src/requirements.txt
codecov==1.6.3
coverage==4.0.3
defusedxml==0.4.1
Django==1.9.2
django-bower==5.1.0
django-cors-headers==1.1.0
django-nose==1.4.3
django-polymorphic==0.9.1
django-vanilla-views==1.0.4
djangorestframework==3.3.3
djangorestframework-xml==1.3.0
ecdsa==0.13
Fabric==1.10.2
#MySQL-python==1.2.5
nose==1.3.7
paramiko==1.16.0
pycrypto==2.6.1
requests==2.9.1
six==1.10.0
psycopg2==2.6.2